+++ mypyc/irbuild/function.py
@@ -30,7 +30,7 @@ from mypyc.sametype import is_same_method_signature
 from mypyc.irbuild.util import concrete_arg_kind, is_constant, add_self_to_env
 from mypyc.irbuild.context import FuncInfo, GeneratorClass, ImplicitClass
 from mypyc.irbuild.statement import transform_try_except
-from mypyc.genops import IRBuilder
+from mypyc.irbuild.builder import IRBuilder
 
 
 # Top-level transform functions
